% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graecolatin.R
\name{greco.latin}
\alias{greco.latin}
\title{graeco latin squares}
\usage{
greco.latin(a, b, c, d, y, i, j)
}
\arguments{
\item{a, b, c, d, y, i, j}{}
}
\value{

}
\description{
In combinatorics, a Graeco-Latin square or Euler square or orthogonal Latin squares of order n over two sets S and T, each consisting of n symbols, is an n×n arrangement of cells, each cell containing an ordered pair (s,t), where s is in S and t is in T, such that every row and every column contains each element of S and each element of T exactly once, and that no two cells contain the same ordered pair.
}
\examples{

}
