% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latinsquare.R
\name{latin}
\alias{latin}
\title{Latin Squares}
\usage{
latin(a, b, c, d, i, j)
}
\arguments{
\item{a, b, c, d, i, j}{}
}
\value{

}
\description{
In combinatorics and in experimental design, a Latin square is an n × n array filled with n different symbols, each occurring exactly once in each row and exactly once in each column.
}
\examples{

}
