% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onequalitative.R
\name{qual}
\alias{qual}
\title{qualitative data (one way)}
\usage{
qual(A, B, Y, r, i, j)
}
\arguments{
\item{A, B, Y, r, i, j}{}
}
\value{

}
\description{
development of concepts which help us to understand social phenomena in natural settings, giving due emphasis to the meanings, experiences and views of the participants
}
\examples{

}
