% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoqualitative.R
\name{qual2}
\alias{qual2}
\title{qualitative data (two way)}
\usage{
qual2(A, B, c, Y, r, i, j)
}
\arguments{
\item{A, B, c, Y, r, i, j}{}
}
\value{

}
\description{
development of concepts which help us to understand social phenomena in natural settings, giving due emphasis to the meanings, experiences and views of the participants
}
\examples{

}
