% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onesplit.R
\name{sp.plot}
\alias{sp.plot}
\title{split(owo way)}
\usage{
sp.plot(block, pplot, splot, Y)
}
\arguments{
\item{block, pplot, splot, Y}{}
}
\value{

}
\description{
split of factor
}
\examples{

}
