% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twosplit.R
\name{sp.plot1}
\alias{sp.plot1}
\title{split(two way)}
\usage{
sp.plot1(block, pplot, splot, aplot, Y)
}
\arguments{
\item{block, pplot, splot, aplot, Y}{}
}
\value{

}
\description{
split of factor
}
\examples{

}
