% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullynested.R
\name{sp.plot2}
\alias{sp.plot2}
\title{Fully Nested}
\usage{
sp.plot2(pplot, splot, aplot, rplot, y)
}
\arguments{
\item{pplot, splot, aploy, rplot, y}{}
}
\value{

}
\description{
If all arguments are stake
}
\examples{

}
