% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosplit.R
\name{sp.plot3}
\alias{sp.plot3}
\title{can not split}
\usage{
sp.plot3(pplot, splot, y)
}
\arguments{
\item{pplot, splot, y}{}
}
\value{

}
\description{
the factor can not split
}
\examples{

}
