% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitsplit.R
\name{ssp.plot}
\alias{ssp.plot}
\title{split split plot}
\usage{
ssp.plot(block, pplot, splot, ssplot, Y)
}
\arguments{
\item{block, pplot, splot, ssplot, Y}{}
}
\value{

}
\description{
split split of factor
}
\examples{

}
